arranged to relate properly to a specific user. Also made it so goals shown are only goals of a particular user, so my goals can't be seen by other users. Left Tags as visible and associatable with other users, So when I create a personal goal, I can link other people's tags to my goal. Might not be an idea we want to keep, but if we want to make users make their own, it isn't a hard change.

+++ src/main/java/cist4830/unomaha/tempo/controllers/TagController.java
@@ -1,6 +1,7 @@
 package cist4830.unomaha.tempo.controllers;
 
 import cist4830.unomaha.tempo.controllers.errors.ResourceNotFoundException;
+import cist4830.unomaha.tempo.model.Goal;
 import cist4830.unomaha.tempo.model.Tag;
 import cist4830.unomaha.tempo.repository.GoalRepository;
 import cist4830.unomaha.tempo.repository.TagRepository;
@@ -9,8 +10,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import cist4830.unomaha.tempo.model.User;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+import cist4830.unomaha.tempo.services.CustomUserPrincipal;
 
 import java.sql.Date;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping(value = "/tags")
@@ -41,7 +47,7 @@ public class TagController {
             , @RequestParam(name = "description") String description) {
         java.util.Date utilDate = new java.util.Date();
         String now = new Date(utilDate.getTime()).toString();
-        Tag tag = new Tag((long) 0, tagstr, description, (long) 1, now, now);
+        Tag tag = new Tag((long) 0, tagstr, description, getLoggedInUser().getId(), now, now);
         tagRepository.create(tag);
         return "redirect:/tags";
     }
@@ -75,8 +81,11 @@ public class TagController {
         Tag tag = tagRepository.findTagById(id).orElseThrow(() -> {
             throw new ResourceNotFoundException();
         });
+        User user = getLoggedInUser();
         model.addAttribute("tag", tag);
-        model.addAttribute("goals", tagRepository.getGoals(tag));
+        model.addAttribute("goals", tagRepository.getGoals(tag).stream()
+            .filter((goal) -> goal.getUserId().equals(user.getId()))
+            .collect(Collectors.toList()));
         return "tags/show";
     }
 
@@ -86,7 +95,18 @@ public class TagController {
             throw new ResourceNotFoundException();
         });
         // ^^ check if it exists first
-        tagRepository.delete(id);
+        // check if user owns
+        User user = getLoggedInUser();
+        if (user.getId() == tag.getUserId())
+            tagRepository.delete(id);
         return "redirect:/tags";
+    }
+
+    public User getLoggedInUser() {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal instanceof CustomUserPrincipal) {
+            return ((CustomUserPrincipal) principal).getUser();
+        }
+        return null;
     }
 }
